+++ app/src/main/assets/CHANGELOG.md
@@ -6,6 +6,11 @@ For support you can use [the contact form](https://contact.faircode.eu/?product=
 
 ### [Ubirajara](https://en.wikipedia.org/wiki/Ubirajara_jubatus)
 
+### Next version
+
+* Added rule action to add local notes
+* Updated [translations](https://crowdin.com/project/open-source-email)
+
 ### 1.2093 - 2023-08-14
 
 * Small improvements and minor bug fixes